refactor: 오디오 연결 로직을 useAudioCommunication 커스텀 훅으로 분리
